requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
argon2-cffi==23.1.0
email-validator==2.1.0.post1
werkzeug==3.0.1
